requirements.txt
pandas>=2.0.0
pydantic>=2.0.0
requests>=2.31.0
tqdm>=4.66.0
openai>=1.0.0
python-dotenv>=1.0.0
wandb>=0.16.0
langchain-openai
langchain-community
langchain-text-splitters
langchain-classic
langchain-core
faiss-cpu
llama-index
llama-index-embeddings-openai
llama-index-llms-openai
nltk>=3.8.0
rouge-score>=0.1.2
bart-score>=0.1.0


